Added background aspects
Added music to the background which I need the lower the volume of. Added two new rooms with more visual aspects for two player or four player options. These rooms also have wall objects that work like the paddles but without recognition of who hit the puck last. The visual and wall changes can be deleted or altered if anyone wishes.